Style: fix UI homepage + add UI Nav Admin + add Button Login/Register

+++ backend/src/main/java/com/bookstore/backend/service/OrdersService.java
@@ -78,7 +78,7 @@ public class OrdersService {
 
 
     // ------------------- GET ORDERS BY USER -------------------
-    public List<OrdersDTO> getOrdersByUser(int userId) {
+    public List<OrdersDTO> getOrdersByUser(Long userId) {
         return ordersRepository.findByUsers_Id(userId)
                 .stream()
                 .map(this::mapToDTO)
